product_rec/requirements.txt
awscli==1.6.2
backports.ssl-match-hostname==3.5.0.1
bcdoc==0.12.2
blinker==1.4
boto==2.33.0
botocore==0.73.0
certifi==2015.4.28
colorama==0.2.5
decorator==3.4.0
docutils==0.12
Flask==0.10.1
Flask-DebugToolbar==0.10.0
Flask-SQLAlchemy==2.1
GraphLab-Create==1.8.1
GraphLab-Create-License==1.8.1
itsdangerous==0.24
Jinja2==2.8
jmespath==0.5.0
MarkupSafe==0.23
mixpanel-py==3.1.1
multipledispatch==0.4.8
prettytable==0.7.2
psycopg2==2.6.1
pyasn1==0.1.9
python-dateutil==2.4.2
requests==2.3.0
rsa==3.1.2
six==1.10.0
SQLAlchemy==1.0.11
sseclient==0.0.8
tornado==4.1
Werkzeug==0.11.3
wheel==0.24.0
